refactor: Add SearchInput component and update imports
